Add MFS, Filestore, Routing, and CAR format APIs
- Implement MFS (Mutable File System) with full 13-method API
- Add FilesApi for MFS operations (cp, ls, mkdir, mv, read, rm, stat, write)
- Add FilestoreApi stubs
- Add IRoutingApi and RoutingApi for content routing
- Add CAR (Content Addressable aRchive) format import/export
- Update IPLD registry and protobuf format

+++ src/LinkedData/ProtobufFormat.cs
@@ -36,7 +36,7 @@ namespace Ipfs.Engine.LinkedData
                 .Select(link => new DagLink(
                     link["Name"].AsString(),
                     Cid.Decode(link["Cid"]["/"].AsString()),
-                    link["Size"].AsNumber().ToInt64Checked()));
+                    (ulong)link["Size"].AsNumber().ToInt64Checked()));
             DagNode node = new(data["data"].GetByteString(), links);
             using MemoryStream ms = new();
             node.Write(ms);
